Remove globals.ts and relocate its contents to appropriate modules
The globals.ts file mixed unrelated concerns: CLI option definitions and
logging configuration. This refactoring improves code organization by moving
each piece to its natural home:

 -  debugOption moved to options.ts, where all other CLI options are defined
 -  configureLogging() moved to log.ts, where recordingSink and other logging
    utilities already reside

This eliminates globals.ts entirely and makes the codebase easier to navigate
by grouping related functionality together.
