+++ packages/cli/src/inbox.tsx
@@ -43,11 +43,10 @@ import process from "node:process";
 import ora from "ora";
 import metadata from "../deno.json" with { type: "json" };
 import { getDocumentLoader } from "./docloader.ts";
-import { configureLogging, debugOption } from "./globals.ts";
-import { tunnelOption } from "./options.ts";
 import type { ActivityEntry } from "./inbox/entry.ts";
 import { ActivityEntryPage, ActivityListPage } from "./inbox/view.tsx";
-import { recordingSink } from "./log.ts";
+import { configureLogging, recordingSink } from "./log.ts";
+import { debugOption, tunnelOption } from "./options.ts";
 import { tableStyle } from "./table.ts";
 import { spawnTemporaryServer, type TemporaryServer } from "./tempserver.ts";
 import { colors, matchesActor } from "./utils.ts";
