+++ packages/cli/src/log.ts
@@ -5,10 +5,10 @@ import {
   type LogRecord,
   type Sink,
 } from "@logtape/logtape";
-import { dirname } from "node:path";
-import process from "node:process";
 import { AsyncLocalStorage } from "node:async_hooks";
 import { mkdir } from "node:fs/promises";
+import { dirname } from "node:path";
+import process from "node:process";
 
 export interface RecordingSink extends Sink {
   startRecording(): void;
@@ -62,3 +62,34 @@ await configure({
   contextLocalStorage: new AsyncLocalStorage(),
   reset: true,
 });
+
+export async function configureLogging() {
+  const logFile = process.env["FEDIFY_LOG_FILE"];
+  await configure({
+    sinks: {
+      console: getConsoleSink(),
+      recording: recordingSink,
+      file: logFile == null ? () => undefined : getFileSink(logFile),
+    },
+    filters: {},
+    loggers: [
+      {
+        category: "fedify",
+        lowestLevel: "debug",
+        sinks: ["console", "recording", "file"],
+      },
+      {
+        category: "localtunnel",
+        lowestLevel: "debug",
+        sinks: ["console", "file"],
+      },
+      {
+        category: ["logtape", "meta"],
+        lowestLevel: "warning",
+        sinks: ["console", "file"],
+      },
+    ],
+    reset: true,
+    contextLocalStorage: new AsyncLocalStorage(),
+  });
+}
